Move Steam Entity to separate file
