+++ homeassistant/components/steam_online/sensor.py
@@ -12,7 +12,6 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import StateType
 from homeassistant.util.dt import utc_from_timestamp
 
-from . import SteamEntity
 from .const import (
     CONF_ACCOUNTS,
     DOMAIN,
@@ -23,6 +22,7 @@ from .const import (
     STEAM_STATUSES,
 )
 from .coordinator import SteamDataUpdateCoordinator
+from .entity import SteamEntity
 
 PARALLEL_UPDATES = 1
